+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Updated
 
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.4.0-alpha.13`
+
 ## [1.0.0-alpha.29] - 2023-01-26
 
 ### Added