+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### Updated
 
+- `gradle -> 8.0.2`
 - `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.4.0-alpha.14`
 
 ## [1.0.0-alpha.29] - 2023-01-26