+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 
 ### Updated
 
-- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.4.0-alpha.15`
+- `com.javiersc.hubdle:com.javiersc.hubdle.gradle.plugin -> 0.4.0-alpha.16`
 - `gradle -> 8.0.2`
 
 ## [1.0.0-alpha.29] - 2023-01-26